refactor: update interfaces to work with Hub API responses
