+++ packages/common/src/search/types/index.ts
@@ -2,7 +2,7 @@ export * from "./ICatalog";
 export * from "./ICollection";
 export * from "./IFacet";
 export * from "./IFacetOption";
-export * from "./types";
-export * from "./IFilter";
-export * from "./IFilterGroup";
 export * from "./IHubApiSearchRequest";
+export * from "./IHubSearchOptions";
+export * from "./IHubSearchResult";
+export * from "./types";
